Add files via upload
Excel spreadsheet to calculate signal to noise ratio for mono (luminance) exposures in astronomy, for stars. Enter the star magnitude, the sky background in mpas (mag per square arc second), the pixel scale in arc seconds, the star FWHM in arc seconds, and camera parameters. It gives you the per exposure stellar SNR and the SNR for a stack of N exposures. It has some empirical constants and the pixel scale has to be below the FWHM to be accurate. Seems to work for my site.